New version of pipeline
The pipeline now runs the newer version of ngs_te_mapper that uses bwa
instead of blat for alignments.
There is also a run script included that takes 6 inputs described in
the read me and runs the whole pipeline.
The test script now downloads the required input files and runs the
main script.